cloudGateway/requirements.txt
confluent_kafka==2.6.1
Flask==3.1.0
PyYAML==6.0.2

